Remove duplicate function names from symbol list
All identifiers with the entity.name.function scope are added to the
local and global symbol lists by default. This commit adds an additional
meta.function.declaration scope only to the first line of a function or
subroutine declaration, and excludes function or subroutine names that
don't have this scope from the symbol lists.